refactor: changed lots of styles to css modules
