+++ src/components/DataDimension/styles/DataTypesSelector.module.css
@@ -1,0 +1,9 @@
+.container {
+    display: flex;
+    flex-flow: column;
+    height: 70;
+    border-bottom: 0;
+    padding-left: 10;
+    padding-right: 5;
+    padding-top: 5;
+}
