--- src/components/DataDimension/styles/DataTypesSelector.style.js
@@ -1,13 +1,0 @@
-export const styles = {
-    container: {
-        display: 'flex',
-        flexFlow: 'column',
-        height: 70,
-        borderBottom: 0,
-        paddingLeft: 10,
-        paddingRight: 5,
-        paddingTop: 5,
-    },
-}
-
-//TODO: Remove unused parts and refactor to css module
