+++ src/components/DataDimension/styles/Details.module.css
@@ -1,0 +1,7 @@
+.detailContainer {
+    display: flex;
+    flex-flow: column;
+    width: 40%;
+    padding-right: 5;
+    padding-left: 5;
+}
