--- src/components/DataDimension/styles/Details.style.js
@@ -1,20 +1,0 @@
-export const styles = {
-    detailContainer: {
-        display: 'flex',
-        flexFlow: 'column',
-        width: '40%',
-        paddingRight: 5,
-        paddingLeft: 5,
-    },
-    titleText: {
-        color: '#616161', // color
-        fontSize: 12,
-        paddingBottom: 10,
-        fontWeight: 300,
-    },
-    dropDown: {
-        padding: 0,
-    },
-}
-
-//TODO: Remove unused parts and refactor to css module
