+++ src/components/FilterField.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import i18n from '@dhis2/d2-i18n'
 
 import Filter from './Filter/Filter'
-import { styles } from './styles/FilterField.style'
+import styles from './styles/FilterField.module.css'
 
 export const FilterField = ({ text, onFilterTextChange, onClearFilter }) => (
     <div style={styles.container}>
