+++ src/components/ItemSelector/styles/ItemSelector.module.css
@@ -1,0 +1,23 @@
+.itemSelectorContainer {
+    box-sizing: border-box;
+    display: flex;
+    flex: 1 1 auto;
+    font-family: Roboto, sans-serif;
+}
+
+.section {
+    border: 1px solid var(--colors-greyLight);
+    display: flex;
+    flex-direction: column;
+    height: 480px;
+    position: relative;
+}
+
+.unselected {
+    margin-right: 55px;
+    width: 418px;
+}
+
+.selected {
+    width: 276px;
+}
