--- src/components/ItemSelector/styles/ItemSelector.style.js
@@ -1,30 +1,0 @@
-import { colors } from './colors'
-import css from 'styled-jsx/css'
-
-export default css`
-    .item-selector-container {
-        box-sizing: border-box;
-        display: flex;
-        flex: 1 1 auto;
-        font-family: Roboto, sans-serif;
-    }
-
-    .section {
-        border: 1px solid ${colors.greyLight};
-        display: flex;
-        flex-direction: column;
-        height: 480px;
-        position: relative;
-    }
-
-    .unselected {
-        margin-right: 55px;
-        width: 418px;
-    }
-
-    .selected {
-        width: 276px;
-    }
-`
-
-//TODO: Remove unused parts and refactor to css module
