+++ src/components/OrgUnitDimension/styles/OrgUnitDimension.module.css
@@ -1,0 +1,6 @@
+.loader {
+    margin-left: auto;
+    margin-right: auto;
+    display: block;
+}
+/* TODO: Can this be removed? the ui-core loader has it's own style to position itself in the center? */
