--- src/components/OrgUnitDimension/styles/OrgUnitDimension.style.js
@@ -1,13 +1,0 @@
-export default {
-    dialogContent: {
-        width: 760,
-        minHeight: 48,
-    },
-    loader: {
-        marginLeft: 'auto',
-        marginRight: 'auto',
-        display: 'block',
-    },
-}
-
-//TODO: Remove unused parts and refactor to css module
