--- src/components/styles/FilterField.style.js
@@ -1,15 +1,0 @@
-import { colors } from '../../modules/colors'
-
-export const styles = {
-    container: {
-        border: `1px solid ${colors.greyLight}`,
-        backgroundColor: colors.white,
-        height: '53px',
-        borderRight: '0px',
-        borderLeft: '0px',
-        borderTop: '0px',
-        padding: '10px 5px 10px 10px',
-    },
-
-    //TODO: Remove unused parts and refactor to css module
-}
